feature-policy/experimental-features/lazyload/lazyload-disabled-tentative.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/feature-policy/experimental-features/resources/common.js"></script>
<style>
html, body {
  height: 100%;
  width: 100%;
}

iframe {
  width: 400px;
  height: 400px;
  margin: 10px;
}

.spacer {
  width: 100%;
  height: 300%;
}
</style>
<div class="spacer"></div>
<script>
  let cross_origin_url =
      "http://{{hosts[alt][www]}}:{{ports[http][0]}}/" +
      "feature-policy/experimental-features/resources/lazyload-contents.html";

  let load_timeout = 100; // ms
  let expected_timeout_msg = false;

  window.scrollTo(0, 0);

  // Sanity-check: Make sure lazyload='on' works as intended.
  promise_test(async(t) => {
    let frame_on = createIframe(document.body, {
        id: "ON",
        lazyload: "on",
        src: `${cross_origin_url}?id=ON`
      });

    // Sanity-check: The frame is not visible.
    assert_greater_than(
        frame_on.getBoundingClientRect().top,
        window.innerHeight * 2,
        "Unexpected position for <iframe> with ID 'ON'.");

    await waitForMessageOrTimeout(t, "ON", load_timeout).then((msg) => {
      assert_equals(msg,
        expected_timeout_msg,
        "With lazyload='on', the frame should not load.");
    });
  }, "Sanity-check: Contents do get lazily loaded when the attribute is 'on'.");

  // Verify that when 'lazyload' policy is disabled,  lazyload='off' and
  // lazyload='auto' behave the same.
  promise_test(async(t) => {
    let frame_on = createIframe(document.body, {
        id: "OFF",
        lazyload: "off",
        src: `${cross_origin_url}?id=OFF`
      });

    // Sanity-check: The frame is not visible.
    assert_greater_than(
        frame_on.getBoundingClientRect().top,
        window.innerHeight * 2,
        "Unexpected position for <iframe> with ID 'OFF'.");

    await waitForMessageOrTimeout(t, "OFF", load_timeout).then(async(msg1) => {
      let message_for_frame_off = msg1;
      // Now do the same for lazyload='auto'.
      let frame_auto = createIframe(document.body, {
        id: "AUTO",
        lazyload: "auto",
        src: `${cross_origin_url}?id=AUTO`
      });

      // Sanity-check: The frame is not visible.
      assert_greater_than(
          frame_on.getBoundingClientRect().top,
          window.innerHeight * 2,
          "Unexpected position for <iframe> with ID 'AUTO'.");

      await waitForMessageOrTimeout(t, "AUTO", load_timeout).then((msg2) => {
        let message_for_frame_auto = msg2;
        assert_equals(
          message_for_frame_off,
          message_for_frame_auto,
          "lazyload='off' not treated as 'auto'.");
      });
    });
  }, "When 'lazyload' feature is disabled, a frame cannot avoid lazyloading " +
     "by setting 'lazyload' attribute to 'off'");
</script>
